requirements.txt
# Зависимости для диаризации аудио
pyannote.audio==3.1.1
torch>=2.0.0,<2.5.0
torchaudio>=2.0.0,<2.5.0
soundfile>=0.12.1
numpy<2.0.0
huggingface_hub<0.20.0
hf-transfer>=0.1.0

# Flask для REST API сервиса
flask>=3.0.0
werkzeug>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
